various componenent name changes + mode navigation added
