+++ client/src/components/Clock.js
@@ -1,45 +1,38 @@
-import React, { useState } from "react";
+import React from 'react';
 import "../css/Clock.css";
+import { secondsToHMS } from '../utils/time';
 
-import Timer from "./Timer";
-import Controls from "./Controls";
-import Indicator from "./Indicator";
+const Clock = ({ elapsed, isBreak, workTime }) => {
 
-const Clock = () => {
-  const [isStarted, setIsStarted] = useState(false);
-  const [elapsed, setElapsed] = useState(0);
-  const [isBreak, setIsBreak] = useState(false);
-  const [workTime, setWorkTime] = useState(0);
-  
+    
+    
+    const { hours, minutes, seconds } = secondsToHMS(isBreak ? (workTime - elapsed) : elapsed);
+    return (
+        <div className="clock d-flex justify-content-between">
+            <div className="clock-value d-flex flex-column align-items-center">
+                <span className="text-light fw-light">Hours</span>
+                <span className="text-light fw-bold display-1">{hours}</span>
+            </div>
 
-  return (
-    <section className="clock d-flex flex-column align-items-center" id="clock">
-      {isBreak ? (
-        <h3 className="text-light">
-          You were able to work for{" "}
-          <span className="text-primary">25m 10s</span> giving you{" "}
-          <span className="text-primary">5m 2s</span> to rest. Keep it up!
-        </h3>
-      ) : (
-        <h3 className="text-light">
-          Time yourself working, your break time is calculated as a{" "}
-          <span className="text-decoration-underline text-primary">ratio</span>{" "}
-          of time worked
-        </h3>
-      )}
+            <div className="d-flex align-items-end">
+                <span className="text-primary fw-bold display-1">.</span>
+            </div>
 
-      <Timer elapsed={elapsed} isBreak={isBreak} workTime={workTime} />
-      <Controls
-        isBreak={isBreak}
-        isStarted={isStarted}
-        hasStarted={elapsed > 0}
-        elapsed={elapsed}
-        workTime={workTime}
-        state={{ setIsStarted, setElapsed, setIsBreak, setWorkTime }}
-      />
-      <Indicator isBreak={isBreak} elapsed={elapsed} workTime={workTime} />
-    </section>
-  );
-};
+            <div className="clock-value d-flex flex-column align-items-center">
+                <span className="text-light fw-light">Minutes</span>
+                <span className="text-light fw-bold display-1">{minutes}</span>
+            </div>
+
+            <div className="d-flex align-items-end">
+                <span className="text-primary fw-bold display-1">.</span>
+            </div>
+            
+            <div className="clock-value d-flex flex-column align-items-center">
+                <span className="text-light fw-light">Seconds</span>
+                <span className="text-light fw-bold display-1">{seconds}</span>
+            </div>
+        </div>
+    );
+}
 
 export default Clock;
