+++ client/src/css/Clock.css
@@ -1,6 +1,5 @@
-
-
 .clock{
-    margin-top: 10vh;
-    position: relative;
+    width: 30%;
+    margin-top: 5vh;
 }
+
