--- client/src/css/Timer.css
@@ -1,5 +1,0 @@
-.timer{
-    width: 30%;
-    margin-top: 5vh;
-}
-
